pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "mosaicraft"
version = "0.1.0"
description = "Perceptual photomosaic generator with Oklab color space, MKL optimal transport, and Laplacian pyramid blending"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "mosaicraft contributors" }]
keywords = [
    "photomosaic",
    "mosaic",
    "image-processing",
    "computer-vision",
    "oklab",
    "perceptual-color",
    "optimal-transport",
    "art",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.23",
    "opencv-python>=4.6",
    "scipy>=1.10",
    "scikit-image>=0.20",
]

[project.optional-dependencies]
faiss = ["faiss-cpu>=1.7"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.5",
    "mypy>=1.5",
    "bandit>=1.7",
    "build>=1.0",
    "twine>=4.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
]

[project.urls]
Homepage = "https://github.com/hinanohart/mosaicraft"
Documentation = "https://github.com/hinanohart/mosaicraft#readme"
Repository = "https://github.com/hinanohart/mosaicraft"
Issues = "https://github.com/hinanohart/mosaicraft/issues"
Changelog = "https://github.com/hinanohart/mosaicraft/blob/main/CHANGELOG.md"

[project.scripts]
mosaicraft = "mosaicraft.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mosaicraft"]

[tool.hatch.build.targets.sdist]
include = [
    "src/mosaicraft",
    "tests",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function call in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]
"examples/*" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-ra", "--strict-markers", "--strict-config"]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/mosaicraft"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
follow_imports = "silent"
allow_untyped_globals = true
check_untyped_defs = false
files = ["src/mosaicraft"]

[[tool.mypy.overrides]]
module = ["cv2", "skimage.*", "scipy.*", "faiss"]
ignore_missing_imports = true
follow_imports = "skip"

[tool.bandit]
exclude_dirs = ["tests", "examples", "benchmarks"]
skips = ["B101"]  # assert_used (used in tests)
